refactor goes brrr
