+++ src/DTC/helpers/iny_codes.h
@@ -1,0 +1,2 @@
+// Lista de codigos DTC para problemas en la inyeccion:
+#define INY_INF
