+++ src/helpers/utilsTimer.c
@@ -1,13 +1,20 @@
-#ifndef UTILSTIMER
-#define UTILSTIMER
+#ifndef UTILSTIMERw
+#define UTILSTIMERw
 
 #include <libopencm3/cm3/nvic.h>
 #include <libopencm3/cm3/systick.h>
-#include "../variables.h"
+#include "../../variables.h"
 /* 
  *Inicia el systick para contar los uS | mS desde que inicio todo
 */
 void utils_timer_setup(void);
+
+/** @brief Loop para funciones que utilizan el systick:
+ */
+void utils_timer_loop(void);
+
+//variables para almacenar los tiempos anteriores para la logica de utils_timer_loop
+uint32_t TimerTriggers[4];
 
 void sys_tick_handler(void){
 	Time.temp++;
@@ -29,5 +36,10 @@ void utils_timer_setup(void){
 	/* Start counting. */
 	systick_counter_enable();
 }
+
+void utils_timer_loop(void){
+	
+}
+
 
 #endif
