+++ src/openefi.c
@@ -7,28 +7,27 @@
 #include <libopencm3/usb/cdc.h>
 // Librerias locales
 #include "defines.h"
-#include "./helpers/utilsTimer.c"
 #include "./usb/usb_conf.c"
 #include "./usb/webusb.c"
 //WIP: Sensores/ADC
-#include "./sensores/utils/input_handler.c"
-
-#define CPWM_ENABLE
-
+#include "./sensores/input_handler.c"
+#include "./helpers/utils.c"
 #ifdef CPWM_ENABLE
 #include "./C_PWM.c"
 #endif
 
 int main(void){
-	int i;
+	// Test Utils dentro de struct & union:
+	utilstimer
+
+	//int i;
 	// setup inicial de clock, IO y USB
 	rcc_clock_setup_in_hse_8mhz_out_72mhz();
-
 	rcc_periph_clock_enable(RCC_GPIOC);
-	gpio_set(GPIOC, GPIO13);
 	gpio_set_mode(GPIOC, GPIO_MODE_OUTPUT_2_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO13);
+	gpio_clear(GPIOC, GPIO13);
 	// iniciamos el systimer asi podemos contar tambien el bootime:
-	utils_timer_setup();
+	//utils_timer_setup();
 	adc_setup(); //WIP: Sensores/ADC
 	usbd_device* usbd_dev = usb_setup();
 	usbd_dev = usb_setup();
@@ -38,12 +37,17 @@ int main(void){
 	#ifdef CPWM_ENABLE
 	c_pwm_setup();
 	#endif
-	gpio_clear(GPIOC, GPIO13);
+	// ---- fin del setup ----
+	gpio_set(GPIOC, GPIO13);
 	while (1){
-		for (i = 0; i < 0x800000; i++){
-			usbd_poll(usbd_dev);
-		}
-		get_input(0); //WIP: Sensores/ADC
+		//utils_timer_loop();
+		usbd_poll(usbd_dev);
+		/*
+			for (i = 0; i < 0x800000; i++){
+				usbd_poll(usbd_dev);
+			}
+			get_input(0); //WIP: Sensores/ADC
+		*/
 		// TODO: resto del ciclo (obtener data de sensores, calculo de tiempos, etc)
 	}
 }
