+++ src/sensors/utils/ema_high_pass.c
