+++ src/sensors/utils/ema_low_pass.c
