--- src/test/usb-poll.cpp
@@ -1,54 +1,0 @@
-#include <libopencm3/stm32/rcc.h>
-#include <libopencm3/stm32/gpio.h>
-#include <libopencm3/cm3/nvic.h>
-#include <libopencm3/stm32/exti.h>
-#include <libopencm3/stm32/timer.h>
-#include <libopencm3/cm3/vector.h>
-#include "C_PWM.cpp"
-#include "dummy.c"
-#include "usb_conf.c"
-#include "winusb.c"
-#include "webusb.c"
-
-/* Set STM32 to 72 MHz. */
-static void clock_setup(void){
-	rcc_clock_setup_in_hse_8mhz_out_72mhz();
-}
-
-static void s_usb(void){
-        {
-            char serial[USB_SERIAL_NUM_LENGTH+1];
-            serial[0] = '\0';
-            target_get_serial_number(serial, USB_SERIAL_NUM_LENGTH);
-            usb_set_serial_number(serial);
-        }
-
-        usbd_device* usbd_dev = usb_setup();
-        webusb_setup(usbd_dev);
-		while (1)
-			usbd_poll(usbd_dev);
-    
-}
-static void gpio_setup(void){
-	/* Enable GPIOC clock. */
-	rcc_periph_clock_enable(RCC_GPIOB);
-	rcc_periph_clock_enable(RCC_GPIOC);
-
-	/* Set GPIO12 (in GPIO port C) to 'output push-pull'. */
-	gpio_set_mode(GPIOB, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO12);
-	gpio_set_mode(GPIOB, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO13);
-	gpio_set_mode(GPIOC, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO13);
-	gpio_set_mode(GPIOB, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO14);
-	gpio_set_mode(GPIOB, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, GPIO15);
-}
-
-int main(void)
-{
-	clock_setup();
-	gpio_setup();
-	tim_setup();
-	exti_setup();
-	s_usb();
-
-	return 0;
-}
